Add local (experimental)
- Use `.nii.gz` instead of `.nii`
- Remove extra `.nii` files.
- Move test into `local/motion_test`
- Replace `extra_stir_dirs.cmake` with `CMakeLists.txt` in `src/local`. Still disabled by default
- Add documentation that if the deformation field is created by giving x-, y-, z-filenames, the images should be in STIR orientation.
