--- examples/nifti/TM.txt
@@ -1,4 +1,0 @@
-0.897104 0.346041 -0.2747 7 
--0.429504 0.828817 -0.358592 -0.613 
-0.103588 0.439679 0.892161 15.871 
-0 0 0 1
